@@ -3,28 +3,20 @@ import 'package:flutter/foundation.dart';
 
 class LifecycleEventHandler extends WidgetsBindingObserver {
   LifecycleEventHandler({
-    this.resumeCallBack,
-    this.suspendingCallBack,
+    required this.resumeCallBack,
   });
 
-  final AsyncCallback resumeCallBack;
-  final AsyncCallback suspendingCallBack;
+  late final AsyncCallback resumeCallBack;
 
   @override
   Future<void> didChangeAppLifecycleState(AppLifecycleState state) async {
     switch (state) {
       case AppLifecycleState.resumed:
-        if (resumeCallBack != null) {
-          await resumeCallBack();
-        }
+        await resumeCallBack();
         break;
       case AppLifecycleState.inactive:
       case AppLifecycleState.paused:
       case AppLifecycleState.detached:
-        if (suspendingCallBack != null) {
-          await suspendingCallBack();
-        }
-        break;
     }
   }
 }
