@@ -20,23 +20,21 @@ typedef PermissionSetCallback = void Function(QRViewController, bool);
 /// and the barcode scanner gets displayed.
 class QRView extends StatefulWidget {
   const QRView({
-    @required Key key,
-    @required this.onQRViewCreated,
+    required Key key,
+    required this.onQRViewCreated,
     this.overlay,
     this.overlayMargin = EdgeInsets.zero,
     this.cameraFacing = CameraFacing.back,
     this.onPermissionSet,
-    this.formatsAllowed,
-  })  : assert(key != null),
-        assert(onQRViewCreated != null),
-        super(key: key);
+    this.formatsAllowed = const <BarcodeFormat>[],
+  }) : super(key: key);
 
   /// [onQRViewCreated] gets called when the view is created
   final QRViewCreatedCallback onQRViewCreated;
 
   /// Use [overlay] to provide an overlay for the view.
   /// This can be used to create a certain scan area.
-  final QrScannerOverlayShape overlay;
+  final QrScannerOverlayShape? overlay;
 
   /// Use [overlayMargin] to provide a margin to [overlay]
   final EdgeInsetsGeometry overlayMargin;
@@ -48,7 +46,7 @@ class QRView extends StatefulWidget {
   final CameraFacing cameraFacing;
 
   /// Calls the provided [onPermissionSet] callback when the permission is set.
-  final PermissionSetCallback onPermissionSet;
+  final PermissionSetCallback? onPermissionSet;
 
   /// Use [formatsAllowed] to specify which formats needs to be scanned.
   final List<BarcodeFormat> formatsAllowed;
@@ -58,21 +56,14 @@ class QRView extends StatefulWidget {
 }
 
 class _QRViewState extends State<QRView> {
-  var _channel;
-  var _observer;
+  late MethodChannel _channel;
+  late LifecycleEventHandler _observer;
 
   @override
   void initState() {
     super.initState();
-    _observer = LifecycleEventHandler(
-        resumeCallBack: () async => {
-              if (_channel != null)
-                {
-                  QRViewController.updateDimensions(widget.key, _channel,
-                      overlay: widget.overlay)
-                }
-            });
-    WidgetsBinding.instance.addObserver(_observer);
+    _observer = LifecycleEventHandler(resumeCallBack: updateDimensions);
+    WidgetsBinding.instance!.addObserver(_observer);
   }
 
   @override
@@ -90,15 +81,17 @@ class _QRViewState extends State<QRView> {
   @override
   void dispose() {
     super.dispose();
-    WidgetsBinding.instance.removeObserver(_observer);
+    WidgetsBinding.instance!.removeObserver(_observer);
+  }
+
+  Future<void> updateDimensions() async {
+    await QRViewController.updateDimensions(
+        widget.key as GlobalKey<State<StatefulWidget>>, _channel,
+        overlay: widget.overlay);
   }
 
   bool onNotification(notification) {
-    Future.microtask(() => {
-          QRViewController.updateDimensions(widget.key, _channel,
-              overlay: widget.overlay)
-        });
-
+    updateDimensions();
     return false;
   }
 
@@ -109,7 +102,7 @@ class _QRViewState extends State<QRView> {
         Container(
           padding: widget.overlayMargin,
           decoration: ShapeDecoration(
-            shape: widget.overlay,
+            shape: widget.overlay!,
           ),
         )
       ],
@@ -139,7 +132,7 @@ class _QRViewState extends State<QRView> {
         break;
       default:
         throw UnsupportedError(
-            "Trying to use the default webview implementation for $defaultTargetPlatform but there isn't a default one");
+            "Trying to use the default qrview implementation for $defaultTargetPlatform but there isn't a default one");
     }
     return _platformQrView;
   }
@@ -149,19 +142,21 @@ class _QRViewState extends State<QRView> {
 
     // Start scan after creation of the view
     final controller = QRViewController._(
-        _channel, widget.key, widget.onPermissionSet, widget.cameraFacing)
-      .._startScan(widget.key, widget.overlay, widget.formatsAllowed);
+        _channel,
+        widget.key as GlobalKey<State<StatefulWidget>>?,
+        widget.onPermissionSet,
+        widget.cameraFacing)
+      .._startScan(widget.key as GlobalKey<State<StatefulWidget>>,
+          widget.overlay, widget.formatsAllowed);
 
     // Initialize the controller for controlling the QRView
-    if (widget.onQRViewCreated != null) {
-      widget.onQRViewCreated(controller);
-    }
+    widget.onQRViewCreated(controller);
   }
 }
 
 class _QrCameraSettings {
   _QrCameraSettings({
-    this.cameraFacing,
+    this.cameraFacing = CameraFacing.unknown,
   });
 
   final CameraFacing cameraFacing;
@@ -174,8 +169,8 @@ class _QrCameraSettings {
 }
 
 class QRViewController {
-  QRViewController._(MethodChannel channel, GlobalKey qrKey,
-      PermissionSetCallback onPermissionSet, CameraFacing cameraFacing)
+  QRViewController._(MethodChannel channel, GlobalKey? qrKey,
+      PermissionSetCallback? onPermissionSet, CameraFacing cameraFacing)
       : _channel = channel,
         _cameraFacing = cameraFacing {
     _channel.setMethodCallHandler((call) async {
@@ -186,9 +181,9 @@ class QRViewController {
             final code = args['code'] as String;
             final rawType = args['type'] as String;
             // Raw bytes are only supported by Android.
-            final rawBytes = args['rawBytes'] as List<int>;
+            final rawBytes = args['rawBytes'] as List<int>?;
             final format = BarcodeTypesExtension.fromString(rawType);
-            if (format != null) {
+            if (format != BarcodeFormat.unknown) {
               final barcode = Barcode(code, format, rawBytes);
               _scanUpdateController.sink.add(barcode);
             } else {
@@ -197,15 +192,14 @@ class QRViewController {
           }
           break;
         case 'onPermissionSet':
-          await getSystemFeatures(); // if we have no permission all features will not be avaible
-          if (call.arguments != null) {
-            if (call.arguments as bool) {
+          if (call.arguments != null && call.arguments is bool) {
+            if (call.arguments) {
               _hasPermissions = true;
             } else {
               _hasPermissions = false;
             }
             if (onPermissionSet != null) {
-              onPermissionSet(this, call.arguments as bool);
+              onPermissionSet(this, call.arguments);
             }
           }
           break;
@@ -220,20 +214,17 @@ class QRViewController {
 
   Stream<Barcode> get scannedDataStream => _scanUpdateController.stream;
 
-  SystemFeatures _features;
-  bool _hasPermissions;
-
-  SystemFeatures get systemFeatures => _features;
+  bool _hasPermissions = false;
   bool get hasPermissions => _hasPermissions;
 
   /// Starts the barcode scanner
-  Future<void> _startScan(GlobalKey key, QrScannerOverlayShape overlay,
-      List<BarcodeFormat> barcodeFormats) async {
+  Future<void> _startScan(GlobalKey key, QrScannerOverlayShape? overlay,
+      List<BarcodeFormat>? barcodeFormats) async {
     // We need to update the dimension before the scan is started.
     try {
       await QRViewController.updateDimensions(key, _channel, overlay: overlay);
       return await _channel.invokeMethod(
-          'startScan', barcodeFormats?.map((e) => e.asInt())?.toList() ?? []);
+          'startScan', barcodeFormats?.map((e) => e.asInt()).toList() ?? []);
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -262,7 +253,7 @@ class QRViewController {
   }
 
   /// Get flashlight status
-  Future<bool> getFlashStatus() async {
+  Future<bool?> getFlashStatus() async {
     try {
       return await _channel.invokeMethod('getFlashInfo');
     } on PlatformException catch (e) {
@@ -273,7 +264,7 @@ class QRViewController {
   /// Toggles the flashlight between available modes
   Future<void> toggleFlash() async {
     try {
-      await _channel.invokeMethod('toggleFlash') as bool;
+      await _channel.invokeMethod('toggleFlash') as bool?;
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -311,7 +302,10 @@ class QRViewController {
     try {
       var features =
           await _channel.invokeMapMethod<String, dynamic>('getSystemFeatures');
-      return SystemFeatures.fromJson(features);
+      if (features != null) {
+        return SystemFeatures.fromJson(features);
+      }
+      throw CameraException('Error', 'Could not get system features');
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -324,12 +318,13 @@ class QRViewController {
   }
 
   /// Updates the view dimensions for iOS.
-  static Future<void> updateDimensions(GlobalKey key, MethodChannel channel,
-      {QrScannerOverlayShape overlay}) async {
+  static Future<bool> updateDimensions(GlobalKey key, MethodChannel channel,
+      {QrScannerOverlayShape? overlay}) async {
     if (defaultTargetPlatform == TargetPlatform.iOS) {
-      // Add small delay to ensure the renderbox is loaded
+      // Add small delay to ensure the render box is loaded
       await Future.delayed(Duration(milliseconds: 100));
-      final RenderBox renderBox = key.currentContext.findRenderObject();
+      if (key.currentContext == null) return false;
+      final renderBox = key.currentContext!.findRenderObject() as RenderBox;
       try {
         await channel.invokeMethod('setDimensions', {
           'width': renderBox.size.width,
@@ -337,9 +332,11 @@ class QRViewController {
           'scanArea': overlay?.cutOutSize ?? 0,
           'scanAreaOffset': overlay?.cutOutBottomOffset ?? 0
         });
+        return true;
       } on PlatformException catch (e) {
         throw CameraException(e.code, e.message);
       }
     }
+    return false;
   }
 }
